ClientApplicationStore
