+++ lib/db/client-applications-store.js
@@ -4,40 +4,13 @@
 const metricsHelper = require('../metrics-helper');
 const { DB_TIME } = require('../events');
 
-const COLUMNS = [
-    'app_name',
-    'created_at',
-    'updated_at',
-    'description',
-    'strategies',
-    'url',
-    'color',
-    'icon',
-];
-const TABLE = 'client_applications';
+const { TABLE, COLUMNS } = require('./utils/const/application-store');
+const {
+    mapRow,
+    remapRow,
+} = require('./utils/mappings/client-application-store');
 
-const mapRow = row => ({
-    appName: row.app_name,
-    createdAt: row.created_at,
-    updatedAt: row.updated_at,
-    description: row.description,
-    strategies: row.strategies,
-    url: row.url,
-    color: row.color,
-    icon: row.icon,
-});
-
-const remapRow = (input, old = {}) => ({
-    app_name: input.appName,
-    updated_at: input.updatedAt,
-    description: input.description || old.description,
-    url: input.url || old.url,
-    color: input.color || old.color,
-    icon: input.icon || old.icon,
-    strategies: JSON.stringify(input.strategies || old.strategies),
-});
-
-class ClientApplicationsDb {
+class ClientApplicationsStore {
     constructor(db, eventBus) {
         this.db = db;
         this.eventBus = eventBus;
@@ -131,4 +104,4 @@ class ClientApplicationsDb {
     }
 }
 
-module.exports = ClientApplicationsDb;
+module.exports = ClientApplicationsStore;
