+++ lib/db/utils/const/application-store.js
@@ -1,0 +1,17 @@
+/* eslint camelcase:off */
+'use strict';
+
+const COLUMNS = [
+    'app_name',
+    'created_at',
+    'updated_at',
+    'description',
+    'strategies',
+    'url',
+    'color',
+    'icon',
+];
+const TABLE = 'client_applications';
+
+module.exports.TABLE = TABLE;
+module.exports.COLUMNS = COLUMNS;
