+++ lib/routes/client-api/register.js
@@ -30,9 +30,7 @@ class RegisterController extends Controller {
             await this.clientApplicationsStore.upsert(clientRegistration);
             await this.clientInstanceStore.insert(clientRegistration);
             this.logger.info(
-                `New client registered with appName=${
-                    clientRegistration.appName
-                } and instanceId=${clientRegistration.instanceId}`
+                `New client registered with appName=${clientRegistration.appName} and instanceId=${clientRegistration.instanceId}`
             );
             return res.status(202).end();
         } catch (err) {
